rename go/team to go/teams
