+++ go/teams/chain_parse.go
@@ -1,4 +1,4 @@
-package team
+package teams
 
 import (
 	"crypto/sha256"
